requirements.txt
asgiref==3.6.0
Django==4.2.1
sqlparse==0.4.4
django-cors-headers==3.1.0
djangorestframework==3.11.0
graphene-django==3.0.0
pyuwsgi==2.0.19.1.post0
requests==2.22.0